+++ lib/rubocop/cli.rb
@@ -101,7 +101,7 @@ module RuboCop
 
     def set_options_to_config_loader
       ConfigLoader.debug = @options[:debug]
-      ConfigLoader.disable_pending_cops = @options[:disable_pending_cops]
+      ConfigLoader.disable_pending_cops = @options[:disable_pending_cops] || @options[:only]
       ConfigLoader.enable_pending_cops = @options[:enable_pending_cops]
       ConfigLoader.ignore_parent_exclusion = @options[:ignore_parent_exclusion]
     end